fix UI

+++ lib/models/review.dart
@@ -4,40 +4,42 @@
 
 import 'dart:convert';
 
-List<Reviews> reviewsFromJson(String str) => List<Reviews>.from(json.decode(str).map((x) => Reviews.fromJson(x)));
+List<Reviews> reviewsFromJson(String str) =>
+    List<Reviews>.from(json.decode(str).map((x) => Reviews.fromJson(x)));
 
-String reviewsToJson(List<Reviews> data) => json.encode(List<dynamic>.from(data.map((x) => x.toJson())));
+String reviewsToJson(List<Reviews> data) =>
+    json.encode(List<dynamic>.from(data.map((x) => x.toJson())));
 
 class Reviews {
-    int bookId;
-    String bookName;
-    String bookAuthor;
-    int bookNumReviews;
-    double bookRating;
-    String bookGenre;
-    String reviewTitle;
-    String review;
-    int ratingNew;
-    int reviewCount;
-    DateTime dateAdded;
-    int reviewPk;
+  int bookId;
+  String bookName;
+  String bookAuthor;
+  int bookNumReviews;
+  double bookRating;
+  String bookGenre;
+  String reviewTitle;
+  String review;
+  double ratingNew;
+  int reviewCount;
+  DateTime dateAdded;
+  int reviewPk;
 
-    Reviews({
-        required this.bookId,
-        required this.bookName,
-        required this.bookAuthor,
-        required this.bookNumReviews,
-        required this.bookRating,
-        required this.bookGenre,
-        required this.reviewTitle,
-        required this.review,
-        required this.ratingNew,
-        required this.reviewCount,
-        required this.dateAdded,
-        required this.reviewPk,
-    });
+  Reviews({
+    required this.bookId,
+    required this.bookName,
+    required this.bookAuthor,
+    required this.bookNumReviews,
+    required this.bookRating,
+    required this.bookGenre,
+    required this.reviewTitle,
+    required this.review,
+    required this.ratingNew,
+    required this.reviewCount,
+    required this.dateAdded,
+    required this.reviewPk,
+  });
 
-    factory Reviews.fromJson(Map<String, dynamic> json) => Reviews(
+  factory Reviews.fromJson(Map<String, dynamic> json) => Reviews(
         bookId: json["book_id"],
         bookName: json["book_name"],
         bookAuthor: json["book_author"],
@@ -50,9 +52,9 @@ class Reviews {
         reviewCount: json["review_count"],
         dateAdded: DateTime.parse(json["date_added"]),
         reviewPk: json["review_pk"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "book_id": bookId,
         "book_name": bookName,
         "book_author": bookAuthor,
@@ -63,7 +65,8 @@ class Reviews {
         "review": review,
         "rating_new": ratingNew,
         "review_count": reviewCount,
-        "date_added": "${dateAdded.year.toString().padLeft(4, '0')}-${dateAdded.month.toString().padLeft(2, '0')}-${dateAdded.day.toString().padLeft(2, '0')}",
+        "date_added":
+            "${dateAdded.year.toString().padLeft(4, '0')}-${dateAdded.month.toString().padLeft(2, '0')}-${dateAdded.day.toString().padLeft(2, '0')}",
         "review_pk": reviewPk,
-    };
+      };
 }
